# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Inventory adjustments dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /adjustments/timeseries:
    get:
      operationId: getAdjustmentsTimeseries
      tags: [adjustments]
      summary: Bucketed inventory adjustment dollars over time
      parameters:
        - $ref: "#/components/parameters/Bucket"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - $ref: "#/components/parameters/StockId"
      responses:
        "200":
          description: Time series
          content:
            application/json:
              schema:
                type: object
                required: [bucket, from, to, points]
                properties:
                  bucket: { $ref: "#/components/schemas/BucketSize" }
                  from: { type: string, format: date }
                  to: { type: string, format: date }
                  points:
                    type: array
                    items: { $ref: "#/components/schemas/TimeseriesPoint" }
  /adjustments/totals:
    get:
      operationId: getAdjustmentsTotals
      tags: [adjustments]
      summary: Total added, removed, and net for a window
      parameters:
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - $ref: "#/components/parameters/StockId"
      responses:
        "200":
          description: Totals
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdjustmentTotals" }
  /adjustments/by-stock:
    get:
      operationId: getAdjustmentsByStock
      tags: [adjustments]
      summary: Adjustments aggregated per Stock ID for a window
      parameters:
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
      responses:
        "200":
          description: Per-stock breakdown
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/StockBreakdownItem" }
  /adjustments/details:
    get:
      operationId: getAdjustmentsDetails
      tags: [adjustments]
      summary: Line-item adjustment rows
      parameters:
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - $ref: "#/components/parameters/StockId"
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 1000, default: 200 }
      responses:
        "200":
          description: Detail rows
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/AdjustmentRow" }
  /adjustments/root-cause:
    get:
      operationId: getAdjustmentsRootCause
      tags: [adjustments]
      summary: Per-stock variance with root-cause investigation tracking
      parameters:
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
      responses:
        "200":
          description: Per-stock variance + investigation
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/RootCauseItem" }
  /adjustments/root-cause/{stockId}:
    put:
      operationId: setVarianceInvestigation
      tags: [adjustments]
      summary: Upsert root-cause investigation for a stock
      parameters:
        - name: stockId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VarianceInvestigationInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VarianceInvestigation" }
...
components:
  parameters:
    Bucket:
      name: bucket
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/BucketSize"
    From:
      name: from
      in: query
      required: true
      description: Inclusive start date (YYYY-MM-DD) parsed from the CC date string
      schema: { type: string, format: date }
    To:
      name: to
      in: query
      required: true
      description: Inclusive end date (YYYY-MM-DD) parsed from the CC date string
      schema: { type: string, format: date }
    StockId:
      name: stockId
      in: query
      required: false
      schema: { type: string }
  schemas:
...
    BucketSize:
      type: string
      enum: [day, week, month, quarter, year]
      default: week
    TimeseriesPoint:
      type: object
      required: [periodStart, periodEnd, label, added, removed, net, addedCount, removedCount]
      properties:
        periodStart: { type: string, format: date }
        periodEnd: { type: string, format: date }
        label: { type: string }
        added: { type: number }
        removed: { type: number }
        net: { type: number }
        addedCount: { type: integer }
        removedCount: { type: integer }
    AdjustmentTotals:
      type: object
      required: [added, removed, net, addedCount, removedCount]
      properties:
        added: { type: number }
        removed: { type: number }
        net: { type: number }
        addedCount: { type: integer }
        removedCount: { type: integer }
    StockBreakdownItem:
      type: object
      required: [stockId, added, removed, net, addedCount, removedCount, onHandValue, onHandFootage]
      properties:
        stockId: { type: string }
        description: { type: string, nullable: true }
        added: { type: number }
        removed: { type: number }
        net: { type: number }
        addedCount: { type: integer }
        removedCount: { type: integer }
        onHandValue:
          type: number
          description: Current on-hand inventory dollar value for this stock.
        onHandFootage:
          type: number
          description: Current on-hand inventory footage for this stock.
        pctOfOnHand:
          type: number
          nullable: true
          description: Net adjustment as a percentage of current on-hand value. Null when on-hand value is zero.
    AdjustmentRow:
      type: object
      required: [id, rollTag, stockId, direction, amount, ccDate, ccString]
      properties:
        id: { type: string }
        rollTag: { type: string }
        stockId: { type: string }
        description: { type: string, nullable: true }
        direction:
          type: string
          enum: [added, removed]
        amount: { type: number }
        ccDate: { type: string, format: date }
        ccString: { type: string }
        poNumber: { type: string, nullable: true }
        usedTikNum: { type: string, nullable: true }
        rowDate: { type: string, format: date, nullable: true }
...
    RootCauseItem:
      type: object
      required:
        [
          stockId,
          addedFootage,
          removedFootage,
          netFootage,
          addedDollars,
          removedDollars,
          netDollars,
          addedCount,
          removedCount,
          onHandValue,
          status,
          rootCauseCategory,
          rootCause,
          investigationOwner,
          correctiveActionStatus,
          correctiveAction,
          correctiveActionOwner,
        ]
      properties:
        stockId: { type: string }
        description: { type: string, nullable: true }
        addedFootage: { type: number }
        removedFootage: { type: number }
        netFootage: { type: number }
        addedDollars: { type: number }
        removedDollars: { type: number }
        netDollars: { type: number }
        addedCount: { type: integer }
        removedCount: { type: integer }
        onHandValue: { type: number }
        pctOfOnHand: { type: number, nullable: true }
        status:
          type: string
          nullable: true
          enum: [open, root_cause_id, closed]
        rootCauseCategory:
          type: string
          nullable: true
          enum:
            [
              missing_from_system,
              missing_from_floor,
              data_error,
              consumed_without_po,
              in_use,
              damage,
              other,
            ]
        rootCause: { type: string, nullable: true }
        investigationOwner: { type: string, nullable: true }
        correctiveActionStatus:
          type: string
          nullable: true
          enum: [not_started, in_progress, complete]
        correctiveAction: { type: string, nullable: true }
        correctiveActionOwner: { type: string, nullable: true }
    VarianceInvestigationInput:
      type: object
      properties:
        status:
          type: string
          nullable: true
          enum: [open, root_cause_id, closed]
        rootCauseCategory:
          type: string
          nullable: true
          enum:
            [
              missing_from_system,
              missing_from_floor,
              data_error,
              consumed_without_po,
              in_use,
              damage,
              other,
            ]
        rootCause: { type: string, nullable: true }
        investigationOwner: { type: string, nullable: true }
        correctiveActionStatus:
          type: string
          nullable: true
          enum: [not_started, in_progress, complete]
        correctiveAction: { type: string, nullable: true }
        correctiveActionOwner: { type: string, nullable: true }
    VarianceInvestigation:
      type: object
      required:
        [
          stockId,
          status,
          rootCauseCategory,
          rootCause,
          investigationOwner,
          correctiveActionStatus,
          correctiveAction,
          correctiveActionOwner,
        ]
      properties:
        stockId: { type: string }
        status:
          type: string
          nullable: true
          enum: [open, root_cause_id, closed]
        rootCauseCategory:
          type: string
          nullable: true
          enum:
            [
              missing_from_system,
              missing_from_floor,
              data_error,
              consumed_without_po,
              in_use,
              damage,
              other,
            ]
        rootCause: { type: string, nullable: true }
        investigationOwner: { type: string, nullable: true }
        correctiveActionStatus:
          type: string
          nullable: true
          enum: [not_started, in_progress, complete]
        correctiveAction: { type: string, nullable: true }
        correctiveActionOwner: { type: string, nullable: true }
...